Add compatibility with newer versions of apcu in apc.php
Was seeing a lot of PHP notices for undefined variables in PHP5.5.18 and above because apcu_cache_info() was returning an array of information that did not match what the code expected. It is expecting a result from apc_cache_info() (note different function name). Added a "compatibility" function that turns newer $cache info keys into older ones so that apc.php will work with both