Readme starts with independent explanation of APCu
The first line of the readme should explain what APCu is.
The old explanation was only sufficient for people who know APC.

The new explanation keeps the comparison to the original APC,
but it explains clearer what the actual purpose of APCu is.

Some new users may find APCu and have never heard of APC.