Print error message if an invalid type is encountered.
Also, remove IS_CALLABLE. It's used for type hints.
It's not used anywhere for `zval` values.
E.g.

```c
Zend/zend_compile.c
4560:                           } else if (arg_info->type_hint ==
IS_CALLABLE && default_ast) {
```

assert() is replaced with a no-op in non-debug binaries.

If there is an unrecognized type, it is probably a symptom of buggy
extension, use of zval after free, or shared memory corruption.
It's useful to have an alert about that.

Instead of assert(0), return the value null and trigger an error.