Fix crash when passing bad array to apcu_delete
If a non refcountable type is a member of that array(e.g. null, int,
float, bool), then APCu would emit a warning, then crash.

Z_TRY_ADDREF will check if something has a reference before attempting
to increment the part of the zval that points to the reference.

This PR includes a test case that would previously cause a segfault.