Make apcu_inc/dec atomic when working on a non-existing entry
We can't just run the updater first and perform a (non-exclusive)
store. This might overwrite the same entry having been inserted
after we released the cache write lock.

Instead add (exclusively) the entry with value 0 and then run the
updater separately.

This also fixes the return value of apcu_cas() on a non-existing
entry, as the return value of the updater was previously ignored
in this case.