Per-entry TTL always has precedence over global TTL
This is part of #291. Previously apcu_store() removed entries those
global TTL has expired, even if a longer per-entry TTL was specified.
Now the per-entry TTL always takes precedence.

Also add a set of helper functions for checking expiry. They are not
yet used everywhere, as the semantics are not consistent...