Try to reduce SMA fragmentation
Instead of picking the first fitting block, check if there is
a smaller block that still fits. We only look at a limited, small
number of additional blocks. This is an approximation of best-fit
allocation.

This is intended to reduce fragmentation, in particular for the
case where a large block was recently freed (and is on top of the
free list) and a small allocation is performed.