Set apc.use_request_time=0 by default
The default used to be 1, which meant that the effective TTL
upon storing and fetching keys was backdated to the beginning
of the web request.

This means values are not stored as long as users might expect,
and might also appear valid longer than expected.

For example, during POST request that 2s and stores a value in
APCu with a TTL of 500ms, the value effectively can't be re-used
by other requests as 0.5 - 2 <= 0.

Closes https://github.com/krakjoe/apcu/issues/391.