Update PHP versions to build
We always use latest PHP revision now.

On PHP 7.4, some tests fail due to spurious output of PHP's built-in
Webserver, which writes some output to stderr which is not caught in
server_test.inc, because the `suppress_output` option of `proc_open()`
does not suppress/ignore stderr, but merely ensures that no Windows
dialog boxes for severe errors are shown, what could cause hangs.  We
solve that by redirecting the childs stderr to a pipe.