Fix bug in serializing php 8.2 lists, reduce memory usage.
Starting in php 8.2, zvals are stored instead of Buckets, halving the
memory usage per extra element. Use the same memory allocation when
unserializing data and when storing data in shared memory in php 8.2.

This was previously copying too much data when there was an associative
array instead of a list, and apc_025.phpt would fail with valgrind.

Related to #323