Change behavior of smart configuration setting
The smart configuration setting can now be used to increase the chance
of full cache wipes during periods of high memory pressure. The idea is
to use smart values > 1 to avoid performance problems when the expunge()
hook is executed too often. The previous behavior was changed because it
could unpredictably cause the insertion of new cache entries to fail,
making the smart setting nearly useless.