Unify shmaddr and header in sma layer
The variables shmaddr and header were often used simultaneously in the
same context, although both always pointed to the same memory address.
This was because the BLOCKAT / OFFSET macros expected shmaddr to be
present and header was used to access data of the header struct.
Now the variable smaheader is used for both.