Drop support for multiple shared memory segments
The support for multiple shared memory segments is not necessary these
days, as systems that suffered from shared memory size limitations now
use mmap(), which is unaffected by such limitations. In addition, this
simplifies the sma layer and makes it easier to refactor the cache layer
so that the cache is relocatable.