Make the key of a cache entry relocatable
The key is now implemented as the last part of apc_cache_entry_t, which
is a step toward making entries relocatable in the future. Furthermore,
this reduces the size of an entry by the size of a pointer and removes
one memcpy() of apc_cache_entry_t during insertion.