Make cache entries relocatable
This makes the persistence representation of cache entries in shm
position-independent by converting all pointers in an entry's zval value
to offsets relative to the entry's starting address. During persistence,
all (process local) pointers are converted and stored as offsets in shm.
During unpersistence, the (process local) shm starting address of the
entry is added to each offset before data access. This is necessary
to make the whole cache relocatable or to develop defragmentation.