Make cache relocatable
This makes the cache position-independent by converting all pointers to
entries into offsets relative to the starting address of the cache
header in the shared memory segment. This is a further step to enable
the use of APCU with processes that access the shared memory segment via
different starting addresses.