Compute the number of hash slots instead of using a fix default value
The new default behavior is to use 512 slots per MB of shared memory,
which keeps the hash table size < 0.4% of shm, which seems negligible.
This should improve the default behavior, as it improves apcu's
performance with many entries, and the hash table size now scales
with the size of shared memory. Previously, a fixed value of 4096 at
32 MB shm was used as default, which corresponds to 128 slots per MB.