Add origin reference to cache entries
Cache entries now contain the address of the pointer/offset used to
access them from a linked list. This can be either the address of the
linked list start pointer or &entry->next of the previous entry.
This is to prepare for defragmentation, as it allows entries to be moved
in memory if the address containing the reference to the current entry
is not available in the context.