Fix Linux CI for PHP 7.x to not silently abort tests (#567)
All PHP 7.x Linux pipelines silently aborted the tests because the PHP
executable could not be found. This was caused by the definition of
"PHP_EXECUTABLE = NONE" in the Makefile, as "php-config --php-binary"
returned "NONE" during ./configure. Since it seems time-consuming to fix
the root cause, this is a temporary workaround until PHP 7.x is removed
from the CI pipeline.