Update TECHNOTES
A large portion of the TECHNOTES no longer reflected the current state
of apcu. The TECHNOTES have therefore been revised to make it easier to
get started. Since the copies of the struct definitions are not required
to provide a general overview, they have been removed to avoid duplicate
maintenance.