Remove all entries that cannot expire during default expunge
When the cache is full, all entries that cannot expire are now removed
by the first cleanup pass. This improves the performance of the default-
expunge operation when apc.ttl = 0 and the application largely does not
use an explicit TTL. Furthermore, this is closer to the behavior of APCu
versions <= 5.1.24. However, unlike APCu <= 5.1.24, the first cleanup
pass still retains all entries that expire via hard- or soft-ttl and
have not yet expired, if apc.ttl = 0.